+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0]
+
+-   Submits the issue on Ctrl+Enter or F1 and then selecting `Submit`.
+-   Uses Monaco's dark theme when using dark theme on GitHub.
+
 ## [0.2.0]
 
 -   Fixes buggy layout when commenting on a source code line.